backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
camelot-py
pillow
sentence-transformers
faiss-cpu
groq
torch
transformers
python-dotenv
pydantic